Replaced the customer data entry file and class. Deleted the old tstCustomer1 and tstCustomer and added a new tstCustomer with tests in place.
