+++ AdminSystem/CustomerDataEntry.aspx.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using ClassLibrary;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
@@ -10,5 +11,30 @@ public partial class _1_DataEntry : System.Web.UI.Page
     protected void Page_Load(object sender, EventArgs e)
     {
 
+    }
+
+   
+    protected void txtCustomerId_TextChanged(object sender, EventArgs e)
+    {
+
+    }
+
+
+
+    protected void txtFullName_TextChanged(object sender, EventArgs e)
+    {
+
+    }
+
+    protected void btnOk_Click(object sender, EventArgs e)
+    {
+        //create a new instance of clsCustomer
+        clsCustomer AnCustomer = new clsCustomer();
+        //capture the fullname
+        AnCustomer.FullName = txtFullName.Text;
+        //st0re the address in the session object
+        Session["AnCustomer"] = AnCustomer;
+        //navigate to the view page
+        Response.Redirect("CustomerViewer.aspx");
     }
 }
